Create the CanvasManager class.
