+++ Assets/Scenes/MainScene.unity
@@ -1193,9 +1193,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 628775c29d3db044b8025dee95e7491d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  pixelPrefab: {fileID: 1372182098702023847, guid: b792750848f1b2247a1a4659efb2731f, type: 3}
   toolTextObject: {fileID: 326312903}
-  moldure: {fileID: 27050991}
+  selectedFormat: 0
 --- !u!4 &525113681
 Transform:
   m_ObjectHideFlags: 0
@@ -4323,6 +4322,52 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1922635731}
   m_CullTransparentMesh: 1
+--- !u!1 &1949177702
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1949177703}
+  - component: {fileID: 1949177704}
+  m_Layer: 0
+  m_Name: Canvas Manager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 2147483647
+  m_IsActive: 1
+--- !u!4 &1949177703
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1949177702}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: -0.17900595, y: 0.016770326, z: -3.6947477}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1949177704
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1949177702}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 57412e93fa20eb84e83683c1fa29cefd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  pixelPrefab: {fileID: 1372182098702023847, guid: b792750848f1b2247a1a4659efb2731f, type: 3}
+  moldure: {fileID: 27050991}
 --- !u!1 &1979016910
 GameObject:
   m_ObjectHideFlags: 0
@@ -5229,6 +5274,7 @@ SceneRoots:
   m_ObjectHideFlags: 0
   m_Roots:
   - {fileID: 525113681}
+  - {fileID: 1949177703}
   - {fileID: 85446812}
   - {fileID: 787059959}
   - {fileID: 586306182}
