+++ Assets/Scripts/CanvasManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 57412e93fa20eb84e83683c1fa29cefd
